feat: refactor financial statement table - replace CollapsibleTable with BaseStatementTable and extract request logic
- remove CollapsibleTable.tsx
- add BaseStatementTable, RequestFunction (fetchRow + data formatting), and CustomRowElement
- update FinancialReportPage to use BaseStatementTable and numeric tab values
- adjust Row to align balances for debit-type elements (asset/expense)
- minor axios client formatting fix
